+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixes
 
+- Fix(Android): duplicated breadcrumbs ([#151](https://github.com/getsentry/sentry-capacitor/pull/151))
 - build(android): Bump sentry-java dependencies to `5.7.0`. ([#154](https://github.com/getsentry/sentry-capacitor/pull/154))
   - [changelog](https://github.com/getsentry/sentry-java/releases/tag/5.7.0)
   - [diff](https://github.com/getsentry/sentry-java/compare/5.0.1...5.7.0)